+++ app/src/androidTest/java/com/battlelancer/seriesguide/ui/ShowsActivityTest.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-// Copyright 2018-2024 Uwe Trottmann
+// Copyright 2018-2025 Uwe Trottmann
 
 package com.battlelancer.seriesguide.ui;
 
@@ -62,10 +62,10 @@ public class ShowsActivityTest {
             e.printStackTrace();
         }
 
-        // Open discover screen
-        ViewInteraction floatingActionButton = onView(
-                allOf(withId(R.id.buttonShowsFloating), isDisplayed()));
-        floatingActionButton.perform(click());
+        // Open search screen
+        ViewInteraction searchButton = onView(
+                allOf(withId(R.id.menu_action_shows_discover_search), isDisplayed()));
+        searchButton.perform(click());
 
         // Added a sleep statement to match the app's execution delay.
         // The recommended way to handle such scenarios is to use Espresso idling resources: